ajout suppression sauce

+++ backend/controllers/sauces.js
@@ -1,4 +1,7 @@
 const Sauce = require('../models/sauces');
+const fs = require('fs');
+
+
 
 // creation de la sauce et enregistrement
 exports.createSauce =  (req, res, next) => {
@@ -44,18 +47,25 @@ exports.modifySauce =  (req, res, next) => {
     });
 };
 
-
-
 // Suppression de la sauce
 exports.deleteSauce = (req, res, next) => {
-
-
-    Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({message: 'Sauce supprimée!'}))
-    .catch(error => res.status(400).json({error}));
-};
-
-
+  Sauce.findOne({ _id: req.params.id})
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                  Sauce.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 // Récupérer une seule sauce
 exports.getOneSauce = (req, res, next) => {
